Search: Make own iframe
Note:
Have yet to include the javascript for search
so nothing calls this at the moment.

The search was positioned in a strange way so have
removed from header.html so it has its own frame which
can be placed anywhere later on. Removed css styling as
not needed now search style is self contained.